+++ core/src/test/resources/org/elasticsearch/index/analysis/test1.json
@@ -6,22 +6,6 @@
                     "type":"standard"
                 }
             },
-            "char_filter":{
-                "my_html":{
-                    "type":"html_strip",
-                    "escaped_tags":["xxx", "yyy"],
-                    "read_ahead":1024
-                },
-                "my_pattern":{
-                    "type":"pattern_replace",
-                    "pattern":"sample(.*)",
-                    "replacement":"replacedSample $1"
-                },
-                "my_mapping":{
-                    "type":"mapping",
-                    "mappings":["ph=>f", "qu=>q"]
-                }
-            },
             "filter":{
                 "stop":{
                     "type":"stop",
@@ -48,21 +32,9 @@
                     "tokenizer":"standard",
                     "filter":["stop", "stop2"]
                 },
-                "custom2":{
-                    "tokenizer":"standard",
-                    "char_filter":["html_strip", "my_html"]
-                },
-                "custom3":{
-                    "tokenizer":"standard",
-                    "char_filter":["my_pattern"]
-                },
                 "custom4":{
                     "tokenizer":"standard",
                     "filter":["my"]
-                },
-                "custom5":{
-                    "tokenizer":"standard",
-                    "char_filter":["my_mapping"]
                 },
                 "custom6":{
                     "tokenizer":"standard",
