+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/CommonAnalysisPlugin.java
@@ -19,6 +19,8 @@
 
 package org.elasticsearch.analysis.common;
 
+import org.elasticsearch.index.analysis.CharFilterFactory;
+import org.elasticsearch.index.analysis.HtmlStripCharFilterFactory;
 import org.elasticsearch.index.analysis.TokenFilterFactory;
 import org.elasticsearch.indices.analysis.AnalysisModule.AnalysisProvider;
 import org.elasticsearch.plugins.AnalysisPlugin;
@@ -26,6 +28,9 @@ import org.elasticsearch.plugins.Plugin;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.TreeMap;
+
+import static org.elasticsearch.plugins.AnalysisPlugin.requriesAnalysisSettings;
 
 public class CommonAnalysisPlugin extends Plugin implements AnalysisPlugin {
     @Override
@@ -34,6 +39,15 @@ public class CommonAnalysisPlugin extends Plugin implements AnalysisPlugin {
         filters.put("asciifolding", ASCIIFoldingTokenFilterFactory::new);
         filters.put("word_delimiter", WordDelimiterTokenFilterFactory::new);
         filters.put("word_delimiter_graph", WordDelimiterGraphTokenFilterFactory::new);
+        return filters;
+    }
+
+    @Override
+    public Map<String, AnalysisProvider<CharFilterFactory>> getCharFilters() {
+        Map<String, AnalysisProvider<CharFilterFactory>> filters = new TreeMap<>();
+        filters.put("html_strip", HtmlStripCharFilterFactory::new);
+        filters.put("pattern_replace", requriesAnalysisSettings(PatternReplaceCharFilterFactory::new));
+        filters.put("mapping", requriesAnalysisSettings(MappingCharFilterFactory::new));
         return filters;
     }
 }
