+++ modules/analysis-common/src/main/java/org/elasticsearch/index/analysis/HtmlStripCharFilterFactory.java
